2025.12.31

+++ Shell.cpp
@@ -1,10 +1,9 @@
 #include "Shell.h"
 
-void Shell::Run(DiskManager &dm, UserManager &um, DirectoryManager &dirm, FileManager &fm, SystemContext &ctx)
+void Shell::Run(DiskManager &dm, UserManager &um, DirectoryManager &dirm, FileManager &fm, LockManager &lm, SystemContext &ctx)
 {
     std::string input;
     std::cout << "欢迎使用FS！ (输入'help'获取指令列表)" << std::endl;
-
     while (true)
     {
         PrintPrompt(ctx, fm);
@@ -14,16 +13,14 @@ void Shell::Run(DiskManager &dm, UserManager &um, DirectoryManager &dirm, FileMa
             continue; // 忽略空输入
         std::vector<std::string> args = ParseInput(input);
         std::string cmd = args[0];
-
         if (cmd == "exit" || cmd == "logout")
         {
             dm.UnMount();
             um.SaveUsersToFile(ctx);
-            // dm.logAction(uname, "exit", "Success");
             std::cout << "再见！" << std::endl;
             break;
         }
-        ExecuteCommand(args, dm, um, dirm, fm, ctx);
+        ExecuteCommand(args, dm, um, dirm, fm, lm, ctx);
     }
 }
 
@@ -34,9 +31,7 @@ std::vector<std::string> Shell::ParseInput(const std::string &input)
     std::stringstream ss(input);
     std::string arg;
     while (ss >> arg)
-    {
         args.push_back(arg);
-    }
     return args;
 }
 
@@ -57,10 +52,9 @@ void Shell::PrintPrompt(SystemContext &ctx, FileManager &fm)
 }
 
 // 执行命令
-void Shell::ExecuteCommand(const std::vector<std::string> &args, DiskManager &dm, UserManager &um, DirectoryManager &dirm, FileManager &fm, SystemContext &ctx)
+void Shell::ExecuteCommand(const std::vector<std::string> &args, DiskManager &dm, UserManager &um, DirectoryManager &dirm, FileManager &fm, LockManager &lm, SystemContext &ctx)
 {
     std::string cmd = args[0];
-
     if (cmd == "help")
     {
         ShowHelp();
@@ -75,7 +69,10 @@ void Shell::ExecuteCommand(const std::vector<std::string> &args, DiskManager &dm
     }
     else if (cmd == "cd")
     {
-        ExecuteCD(args[1], fm);
+        if (args.size() < 2)
+            std::cout << "用法: cd <dirname>" << std::endl;
+        else
+            ExecuteCD(args[1], fm);
     }
     else if (cmd == "mkdir")
     {
@@ -96,21 +93,32 @@ void Shell::ExecuteCommand(const std::vector<std::string> &args, DiskManager &dm
         if (args.size() < 2)
             std::cout << "用法: rm <filename>" << std::endl;
         else
-            ExecuteRM(args[1], dirm, fm, &dm);
+            ExecuteRM(args[1], dirm, fm, &dm, lm);
     }
     else if (cmd == "cat")
     {
         if (args.size() < 2)
             std::cout << "用法: cat <filename>" << std::endl;
         else
-            std::cout << fm.ReadFile(args[1]) << std::endl;
+            ExecuteCat(args[1], dirm, lm, fm);
     }
     else if (cmd == "write")
     {
         if (args.size() < 3)
             std::cout << "用法: write <filename> <content>" << std::endl;
         else
-            fm.WriteFile(args[1], args[2]);
+        {
+            std::string full_content = "";
+            for (size_t i = 2; i < args.size(); ++i)
+            {
+                full_content += args[i];
+                if (i != args.size() - 1)
+                {
+                    full_content += " ";
+                }
+            }
+            ExecuteWrite(args[1], full_content, dirm, lm, fm);
+        }
     }
     else
     {
@@ -123,6 +131,7 @@ void Shell::ShowHelp()
 {
     std::cout << "支持的指令:\n"
               << "  ls                  列出目录内容\n"
+              << "  cd    <目录名>      切换当前工作目录\n"
               << "  mkdir <名称>        创建目录\n"
               << "  touch <名称>        创建空文件\n"
               << "  rm    <名称>        删除文件或目录\n"
@@ -162,7 +171,7 @@ void Shell::ExecuteCD(const std::string &path, FileManager &fm)
 }
 
 // 执行删除文件逻辑
-void Shell::ExecuteRM(const std::string &filename, DirectoryManager &dirm, FileManager &fm, DiskManager *disk)
+void Shell::ExecuteRM(const std::string &filename, DirectoryManager &dirm, FileManager &fm, DiskManager *disk, LockManager &lm)
 {
     // 1. 安全检查：获取该文件的 Inode
     uint32_t inodeId = dirm.FindInodeId(filename, fm.GetCurrentInodeId());
@@ -171,21 +180,70 @@ void Shell::ExecuteRM(const std::string &filename, DirectoryManager &dirm, FileM
         std::cerr << "错误: 无法删除 '" << filename << "': 没有那个文件或目录！" << std::endl;
         return;
     }
-    Inode node;
-    if (disk->ReadInode(inodeId, node))
-        // 2. 检查是否为目录 (mode == 2)
-        if (node.mode == 2)
-        {
-            std::cerr << "rm: 无法删除 '" << filename << "': 是一个目录" << std::endl;
-            return;
-        }
+    // 2. 申请写权限
+    if (!lm.RequestAccess(inodeId, true))
+    {
+        std::cerr << "文件保护：无法删除 '" << filename << "'正在被其他用户访问，请稍后再试!" << std::endl;
+        return;
+    }
     // 3. 执行删除
     if (fm.DeleteFile(filename))
     { // 删除成功，模拟 Linux 默认不输出
     }
     else
-    {
         std::cerr << "rm: 删除失败" << std::endl;
+    // 无论删除成功与否，操作结束必须释放锁
+    lm.ReleaseAccess(inodeId, true);
+}
+
+// 执行显示文件内容逻辑
+void Shell::ExecuteCat(const std::string &filename, DirectoryManager &dirm, LockManager &lm, FileManager &fm)
+{
+    // 1. 获取文件的 Inode ID (必须先知道 ID 才能锁住对应的资源)
+    uint32_t inodeId = dirm.FindInodeId(filename, fm.GetCurrentInodeId());
+    if (inodeId == (uint32_t)-1)
+    {
+        std::cout << "错误：文件 '" << filename << "' 不存在!" << std::endl;
         return;
+    }
+    else
+    {
+        // 2. 申请“读权限” (isWrite = false)
+        if (lm.RequestAccess(inodeId, false))
+        {
+            std::this_thread::sleep_for(std::chrono::seconds(10)); // 模拟读大文件耗时
+            // 3. 执行读取操作
+            std::cout << fm.ReadFile(filename) << std::endl;
+            // 4. 读取完成后，必须释放读权限
+            lm.ReleaseAccess(inodeId, false);
+        }
+        else
+            // 如果返回 false，说明此时有人正在“写”或者“删除”该文件
+            std::cout << "文件保护：文件 '" << filename << "' 正在被其他用户访问，请稍后再试!" << std::endl;
+    }
+}
+
+void Shell::ExecuteWrite(const std::string &filename, const std::string &content, DirectoryManager &dirm, LockManager &lm, FileManager &fm)
+{
+    // 1. 查找 Inode ID
+    uint32_t inodeId = dirm.FindInodeId(filename, fm.GetCurrentInodeId());
+    if (inodeId == (uint32_t)-1)
+    {
+        std::cout << "错误：文件 '" << filename << "' 不存在!" << std::endl;
+        return;
+    }
+    else
+    {
+        // 2. 申请写权限
+        if (lm.RequestAccess(inodeId, true))
+        {
+            // 3. 执行实际的写入操作
+            fm.WriteFile(filename, content);
+            // 4. 操作完成后必须释放权限
+            lm.ReleaseAccess(inodeId, true);
+        }
+        else
+            // 如果 RequestAccess 返回 false，说明有人正在 cat (读) 或正在 write (写)
+            std::cout << "文件保护：文件 '" << filename << "' 正在被其他用户访问，请稍后再试!" << std::endl;
     }
 }
